CAF-3647: Moved queue services to separate project which can be used by both Job Service and Job Tracking Worker.
