+++ deploy/src/main/resources-filtered/docker-compose.yml
@@ -52,6 +52,9 @@ services:
     env_file:
       - ./rabbitmq.env
     environment:
+      CAF_STATUS_CHECK_TIME: 5
+      CAF_TRACKING_PIPE: jobtracking-in
+      CAF_WEBSERVICE_URL: http://jobservice:8080/job-service/v1
       CAF_WORKER_INPUT_QUEUE: jobtracking-in
       CAF_WORKER_OUTPUT_QUEUE: jobtracking-out
       JOB_DATABASE_URL: jdbc:postgresql://jobservicedb:5432/jobservice
