+++ job-service-queue-services/pom.xml
@@ -1,0 +1,67 @@
+<!--
+
+    Copyright 2015-2017 EntIT Software LLC, a Micro Focus company.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+         http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+-->
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <artifactId>job-service-queue-services</artifactId>
+  <packaging>jar</packaging>
+  <name>job-service-queue-services</name>
+
+  <parent>
+    <groupId>com.github.jobservice</groupId>
+    <artifactId>job-service-aggregator</artifactId>
+    <version>2.3.0-SNAPSHOT</version>
+  </parent>
+
+  <dependencies>
+    <!-- Codec serialization. -->
+    <dependency>
+      <groupId>com.github.cafapi.util</groupId>
+      <artifactId>util-moduleloader</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.github.cafapi.codec</groupId>
+      <artifactId>codec-json</artifactId>
+      <scope>runtime</scope>
+    </dependency>
+
+    <!-- RabbitMQ queuing services support -->
+    <dependency>
+      <groupId>com.github.cafapi</groupId>
+      <artifactId>caf-api</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.github.workerframework</groupId>
+      <artifactId>worker-configs</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.github.workerframework</groupId>
+      <artifactId>worker-queue-rabbit</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.github.workerframework</groupId>
+      <artifactId>util-rabbitmq</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>commons-codec</groupId>
+      <artifactId>commons-codec</artifactId>
+    </dependency>
+  </dependencies>
+
+</project>
