+++ job-service-queue-services/src/main/java/com/hpe/caf/services/job/queue/services/configuration/QueueServicesConfig.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2015-2017 EntIT Software LLC, a Micro Focus company.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.hpe.caf.services.job.queue.services.configuration;
+
+/**
+ * Configuration class for RabbitMQ properties.
+ */
+public class QueueServicesConfig {
+
+    public static String getRabbitMQHost(){
+        return System.getenv("CAF_RABBITMQ_HOST");
+    }
+
+    public static int getRabbitMQPort(){
+        return Integer.parseInt(System.getenv("CAF_RABBITMQ_PORT"));
+    }
+
+    public static String getRabbitMQUsername(){
+        return System.getenv("CAF_RABBITMQ_USERNAME");
+    }
+
+    public static String getRabbitMQPassword(){
+        return System.getenv("CAF_RABBITMQ_PASSWORD");
+    }
+
+}
