+++ job-service/pom.xml
@@ -238,6 +238,11 @@
       <version>1.2</version>
     </dependency>
     <!-- End Logging -->
+    <dependency>
+      <groupId>com.github.jobservice</groupId>
+      <artifactId>job-service-queue-services</artifactId>
+      <version>${project.version}</version>
+    </dependency>
 
   </dependencies>
