+++ pom.xml
@@ -68,6 +68,7 @@
         <module>job-service-caller</module>
         <module>job-service-ui</module>
         <module>job-service-contract</module>
+        <module>job-service-queue-services</module>
         <module>job-service</module>
         <module>job-service-container</module>
         <module>job-service-client</module>
