+++ worker-jobtracking/pom.xml
@@ -101,6 +101,11 @@
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
         </dependency>
+        <dependency>
+            <groupId>com.github.jobservice</groupId>
+            <artifactId>job-service-queue-services</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 
 </project>
